make the ByteXXXStream classes inner classes of the TextExtractors that use them.
