+++ KeywordSearch/src/org/sleuthkit/autopsy/keywordsearch/FileTextExtractor.java
@@ -18,9 +18,17 @@
  */
 package org.sleuthkit.autopsy.keywordsearch;
 
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.Reader;
 import java.util.Arrays;
 import java.util.List;
 import org.apache.solr.common.util.ContentStream;
+import org.openide.util.NbBundle;
+import org.sleuthkit.autopsy.coreutils.Logger;
+import org.sleuthkit.datamodel.AbstractContent;
 import org.sleuthkit.datamodel.AbstractFile;
 
 /**
@@ -121,4 +129,70 @@ abstract class FileTextExtractor<AppendixProvider> extends TextExtractor<Appendi
     String getName(AbstractFile source) {
         return source.getName();
     }
+
+    private static class ByteContentStream implements ContentStream {
+
+        //input
+        private final byte[] content; //extracted subcontent
+        private long contentSize;
+        private final AbstractContent aContent; //origin
+
+        private final InputStream stream;
+
+        private static final Logger logger = Logger.getLogger(ByteContentStream.class.getName());
+
+        public ByteContentStream(byte[] content, long contentSize, AbstractContent aContent) {
+            this.content = content;
+            this.aContent = aContent;
+            stream = new ByteArrayInputStream(content, 0, (int) contentSize);
+        }
+
+        public byte[] getByteContent() {
+            return content;
+        }
+
+        public AbstractContent getSourceContent() {
+            return aContent;
+        }
+
+        @Override
+        public String getContentType() {
+            return "text/plain;charset=" + Server.DEFAULT_INDEXED_TEXT_CHARSET.name(); //NON-NLS
+        }
+
+        @Override
+        public String getName() {
+            return aContent.getName();
+        }
+
+        @Override
+        public Reader getReader() throws IOException {
+            return new InputStreamReader(stream);
+
+        }
+
+        @Override
+        public Long getSize() {
+            return contentSize;
+        }
+
+        @Override
+        public String getSourceInfo() {
+            return NbBundle.getMessage(this.getClass(), "ByteContentStream.getSrcInfo.text", aContent.getId());
+        }
+
+        @Override
+        public InputStream getStream() throws IOException {
+            return stream;
+        }
+
+        @Override
+        protected void finalize() throws Throwable {
+            super.finalize();
+
+            stream.close();
+        }
+
+    }
+
 }
